Resolves #66: Fix bad usage of StringRef in dropIndexActor(), causing memory corruption.
